refactor: split Copilot runtime helpers
