+++ src/agents/embedded-agent-runner/compact.hooks.harness.ts
@@ -643,7 +643,6 @@ export async function loadCompactHooksHarness(): Promise<{
   vi.doMock("../harness/policy.js", () => ({
     resolveAgentHarnessPolicy: resolveAgentHarnessPolicyMock,
   }));
-
   vi.doMock("../harness/runtime-plugin.js", () => ({
     ensureSelectedAgentHarnessPlugin: vi.fn(async () => undefined),
   }));
