+++ src/agents/embedded-agent-runner/run.overflow-compaction.harness.ts
@@ -678,7 +678,6 @@ export async function loadRunOverflowCompactionHarness(): Promise<{
     prepareProviderExtraParams: vi.fn(async () => ({})),
     wrapProviderStreamFn: vi.fn((_cfg: unknown, _model: unknown, fn: unknown) => fn),
   }));
-
   vi.doMock("../auth-profiles.js", () => ({
     isProfileInCooldown: mockedIsProfileInCooldown,
     markAuthProfileFailure: mockedMarkAuthProfileFailure,
